Optimization
Replaced the tryCombinations function with a stub that can invoke either a fast_fit knapsack based solver or a very slow brute-force solver for validating results. Default is to use the knapsack method. The current getBestTrade also culls the item list of items that are worth less than the cheapest item. This may be a premature optimization.