Presentation cleanup
Normalized str() and name() across station and system etc but I don't know that I like what they do; I think str() should probably be the fancy version (so that print(station) does the right thing) and name() returns just the dbname. But if I do that, there's not so much point in hiding dbname, is there?