Optimize building of links data by using squared values instead of roots.
This requires the consumer to perform their own math.sqrt() but it reduces the number of them done during loading, etc, and in most cases you can simply compare the square of your desired distance against the stored value. If you are limiting to a max of 5 lightyears, then you are limiting to a max of 5^2 lightyears^2.