'TradeGitty': NOT READY FOR PRODUCTION
More work on the big refactor to separate concerns.

This version introduces the commands subdirectory (to replace subcommands) and replaces 'TradeEnv' with 'CommandEnv',
and it breaks out the functionality in the command modules into a 'run()' and 'render()' state.

(I might rename 'run' to 'query' or something)

A big part of the change is that we only load stuff for the current command unless you are using -h
or have generated a usage request somehow.

Should speed up loading a lot but this should also put us in a better standing to do lazy loading
etc.

One problem I don't yet feel I've solved is how we're going to re-use the TradeDB between calls;
it's fine if you are just changing a few outlying parameters, but what if you do something
like change the db path?