Issue #50 Interaction between -0 and demand for a sold item
Try to honor -0 a little more while trying to avoid the risk of users making items inaccessible.

This change means that -0 will be honored if the "fromStn" price is 0, but if the station is
buying an item (fromStn > 0) then it will revert to using the default (unk) for the demand
field.