TradeDB.genSystemsInRange
As the database grows, the cost of populating System.links for every
star is going to become more and more significant in memory and cpu.

This functionality provides an on-demand spatial lookup with caching;
most of the time we're probably going to be interested in stars within,
say 60ly of Aulin.

Usage:

```
    import tradedb, math
    tdb = tradedb.TradeDB(buildLinks=False, includeTrades=False)
    aulin = tdb.lookupSystem("Aulin")
    for dst, distSq in tdb.genSystemsInRange(aulin, ly=8.0):
        print("{:<30} {:>6.2f}".format(dst.name(), math.sqrt(distSq)))
```

TODO: Remove System.links and use this instead.