normalizedStr adjustments
In order to increase the much larger namespace, we're going to require users to sometimes be more specific in naming things. To support this, I've made the normalizer a little less aggressive. This should be followed by some changes which make the lookup system a little more aggressive and also explain in more detail what the ambiguities that arise are.