Initial pass at "lookupPlace"
LookupPlace should replace "lookupSystem" and "lookupStation" because you're frequently going to wind up needing qualification.

It can accept a system or station name, or a string made of a "system/station" and use partials and ranked matching to try and do its best to find a singular match.

the "system/station" lookup is not implemented in this pass.