Fix for range calculations and accuracy of nav
This allows us to be more rigorous in finding the shortest path without using a ton more cpu; it may decreases performance of some short routes but it should amortize (average performance will be improved).

It also now reports the correct numbers for jump, total and direct.

Thanks to bgol for catching that the outputs were wrong in the first place.