Automatically add placeholders for stations
While parsing TradeDangerous.prices or a .prices import, if the
"--ignore-unknown" (-i) flag is specified, instead of just ignoring
unrecognized stations, we now add a placeholder for it to the cache.

Note that when you rebuild the cache, the placeholders will be lost,
so you will need to buildcache -i.