Sqrt is not the enemy, math.sqrt() is the enemy.
I realized that it's the function call part of math.sqrt not the actual root that makes it so expensive. Timing suggests that calling "math.sqrt" costs ~300ns on a 1st gen i7 vs 29ns to do "** 0.5", so I introduced a "Station.distanceTo" call and eliminated all calls to math.sqrt in favor of ** 0.5.