Integrate EDDBlink changes into TD, remove change code from plugin.
The changes previously made by the EDDBlink plugin to the database and
csvexport.py have been merged into those files, so the plugin no longer
needs to check for those changes or make them if they haven't happened.

As such, that code has been removed from the plugin.

EDDBlink plugin previously checked if it was the first run by looking
for one of those changes. It now checks to see if the
TradeDangerous.prices file exists. The plugin WILL run a 'clean' import
if it sees that the prices file has not been created, fair warning.

Since most of the data will now be updated automatically by the plugin,
those .csv files have been added to gitignore and removed from the repo.