Added recursionLimitedFit and simpleFit fit functions.
recursionLimitedFit is precisely the same as fastFit, bnut with a
recursion limiter.

simpleFit is as described on the forum thread.

The line that sets which fitFunction to use:
"self.defaultFit = fit or self.fastFit"

has been commented out and duplicated with the two new functions to make
it a bit easier to switch between them during testing.

Will probably add a command-line switch at some point.