Fix import of Krait Mk II
So, there is disagreement.

Coriolis has Krait Mk II
EDDB has Krait MkII

Unlike other similar ships it's not Mk.  (see the '.') either, so completely different to other cases.

(In game it appears to actually be Krait MkII, so EDDB looks to be correct and maybe we blame FDEV for being inconsistent as the root of the problem)

So we have this fix until/unless things change.