Add "tdh" option to edapi plugin for TDHelper to use.
Trade Dangerous Helper uses the EDAPI response to get information about
the CMDR, such as amount of credits, owned ships, etc.

Right now, it's using the edce client to get that response, but it can
just as easily use the EDAPI plugin included with TD.

The EDAPI plugin has a save option which outputs the EDAPI response to
"<TD Install>/tmp/profile.<current_date>_<current_time>.json", eg.
"profile.20181028_091544.json".

This has a couple drawbacks, one being that every time the plugin is run
with the save option, it creates a new file, which just means more
things for the client to take care of, and TDH would need to figure out
what the file is named everytime.

To circumvent both, a new option has been added to the plugin that will
cause the plugin to always save to "<TD Install>/tmp/tdh_profile.json".

The original "save" option is unaffected.

The command to generate the TDH file is:
"trade.py import -P edapi -O tdh"