Remove old tdh_profile before writing new, quit after save if TDH.
If the plugin was called specifically to provide TDH with the EDAPI
response (i.e., by using the "tdh" option), nothing needs to done once
the file has been saved, so exit immediately.

Also, because the plugin will be saving to the same filename every time,
delete the old file before saving a new one to prevent any weirdness
occurring.