Not implemented plugin finish() raises Warning, not Error, returns True
The finish() method should never be required to be implemented.

If the plugin needs to do something at the point where finish() is
called by the import command, then it's nice to have the option, but if
the plugin doesn't need to do anything there, it shouldn't be required
to implement the method anyway just to say "Nope, I'm good, get on with
it." and quit.

So, if the method is not implemented, instead of throwing an Exception,
it now throws a Warning, and returns True by default, True being the
"Nope, I'm good, get on with it."