Future-proofing, hopefully.
EDDBlink plugin updated to include newest items added in E:D,
temporarily using their fdev_id as their item_id.

EDDBlink plugin updated to detect when an item's item_id has changed
(such as when an item that was not in EDDB.io's commodity.json and was
using the fdev_id has been added to the json and therefore now has an
item_id (which definitely is not the fdev_id)) and update the database
to use the new item_id, using the item's fdev_id to check.

corrections.py updated to correctly mark the 'OCCUPIED CRYOPOD' as
'Occupied Escape Pod' and not DELETED.

SQL script updated to make fdev_id in the Item table unique and and an
index using the fdev_id as key. (Does not take effect until a 'clean'
run of the plugin is performed.)