Generalize faulty EDDB.io entry check.
Any item that doesn't have an ed_id value is 'faulty', so skip importing
any of them.

As of right now, that's just #270 "Occupied Cryo Pod', but thjat may
change in the future.

("OccupiedCryoPod" is the symbol for "Occupied Escape Pod".)

Also updated the manual corrections to include the escape pod.