Fix #31.
Category.csv is deleted on a clean run, so it can't be used in that
case.

Added a default value of what the categories currently are to use when
the file doesn't exist.

It's not likely that the categories will ever change, but this is for in
case they do.