Update listings import to use faster method, various fixes.
Removed some old code that isn't needed anymore, added some code to make
the import a bit more stable and resistant to borking the database if
the program is exited early.

Changed the listings import method to use executemany instead of doing
each listings import as a separate transaction, resulting in a huge
speedup on the typical case and a pretty decent speed up in general.