refactor: Add py3.7 to version list, build lowest compat., init TD ver.
Lead developer is using 3.7, so obviously tests need to include that
version.

If we're only going to build for one python version, it should be the
lowest compatible version.

Last official version of TD was 7.4.0, back in 2016-01-29, so
considering how much has changed since then, should start at something
more like 10.

The .1 part of 10.0.1 is this specific refactor.