refactor: remove template folder environment variable
The template folder contains the .sql file that is needed to build TD's
database.

TD copies these files whenever it needs to create a new database.

It is vital that these files exist and be correctly pointed to in order
for TD to work.

It makes absolutely no sense to make it possible for the user to even
accidentally point TD at a non-existant path or even an existing one
that simply doesn't have those files, and there is no need to have these
template files for any purpose other than that of TD using them to
create a new database.

In summary:
having this environment variable:
  usefulness: <=0
  potential harm: >0