feat: automatically update Added.csv, RareItem.csv, TradeDangerous.sql
refactor: restore whitespace to tradedb.py

The template files may need updating, and if that is the case, any TD
database will need their copy updated as well.

Since TD detects changes to the 'cache', any updates to these files will
be integrated into TD the next time it is run.

Changes to the SQL will likely require doing a 'clean' run with the
eddblink plugin as they are often breaking changes.