feat: Add environment variable for setting location of 'eddb' folder.
If set, the "dump files" will be stored to the path set by TD_EDDB. It
will not be in an 'eddb' sub-folder, but the specific path set by the
environment variable.

(listings.csv => $TD_EDDB\listings.csv)

If not, the "dump files" will be stored in the 'data' folder (which
itself may be set via TD_DATA) in an 'eddb' sub-folder.

(listings.csv => $TD_DATA\eddb\listings.csv)