fix: Don't crash when attempting to download non-existent file.
Fixes #21.

Specifically, it fixes the crash that happens because of the ship's
index file no longer existing of the EDCD coriolois data github.

So the Ships database won't be getting updated until a new source is
found to download from.

For now the latest version available before the removal is stored as a
template.

TD will get another update once we have a new site from which to get the
index again.