beta: Re-implement everything done before API switch.
update command had a mutually exclusive group consisting of one argument
and another mutually existing group consisting of four other arguments.
All five arguments are all mutually exclusive, so I removed the group of
groups and put them all in the same mutually exclusive group.

buildcache had the only instance of the shortcut being listed before the
command ( "'-f', '--force'" rather than "'--force', '-f'"), so that's
been fixed to conform with the normal.

We now have a dict of each option, for each command, which is about
where we were before the switch to appJar.