beta: Some work populating required and common args.
The required arguments section is finished, in terms of existing and
functioning correctly.

Whilst working on the debug and detail common options, I noticed that
the selectors were acting backwards, so I had to do a bit of a library
overwrite to fix that weirdness.

If you test this, let me know if the number increases when pressing the
up arrow and decreases when pressing the down arrow. This is the
intended behaviour that is caused by the overwrite.

The normal behaviour is that increase is down arrow, and decrease is up
arrow, at least for me on my machine.

You can check out the difference by (un)commenting the following two
lines:
```
80: gui._populateSpinBox = _populateSpinBox
81: gui.setSpinBoxPos = setSpinBoxPos
```