beta: Add remaining common arguments to window.
This is not the final design, this will probably change eventually.
Right now I'm focusing purely on functionality and space. When
functionality is complete, visuals will get a look.

Regarding functionality, all that remains of the basics is populating
the optional arguments, enabling the running of user-chosen TD commands,
and redirecting TD's output to the window.
Upon finishing those bits, TD GUI will be precisely as functional as TD
CLI, and possibly a bit easier to use, though not nearly as user
friendly as TDHelper.
This point will mark the first public release of the GUI to the main
branch, although it will still be considered in Beta status.

There's more planned after that point.