beta: Store MutuallyExclusiveGroup data to optArg dict.
Members of a MutuallyExclusiveGroup now have the names of the other
members of the group stored as a list under 'excludes'.

This will be used to reset arguments to default/off when a different
arguments excludes it.

Worked out the function design of the various kinds of arguments, to be
implemented in the future.

Fixed '--credit' and '--capacity' required arguments for the 'run'
command to show they are numbers. Probably need to make a verification
method for them as well as the ints and 'credits's floating in the
optional arguments.