beta: Major reformat to argument dictionary, now includes widget type
The former reqArg and optArg dicts have been merged along with the
common arguments into the single allArgs dict.

The dict formerly known as allArgs is now known as argVals to accomodate
the above as well as more accurately reflect its purpose.

During the creation of the dict, each argument will have the appropriate
widget information for the to-be-completed makeWidget method.

During population of the arguments panels, makeWidget will be used to
perform the actual population. Currently only arguments that take a
string are implemented as a proof-of-concept.

When makeWidget is completed, the common arguments will added to the
window using makeWidget as well.

Current TODO list:

1) Finish implementing makeWidget and use it for populating /all/
arguments, including top-level. The '--option' argument will be treated
as a string for now.

2) Implement runTD, which includes grabbing the output and redirecting
it to the GUI. At this point the help panel will be adjusted to be a
tabbed display, so that the command output can be on a separate tab.
(Create new tabs as needed, close button for user?)

---- Initial beta release ----

3) Fully implement the '--option' argument, possibly by having an entry
containing the current option list, and a button that will open a sub
window containing all the options for the chosen plugin.

4) Fix discovered bugs.

5) Implement beta tester feedback until no longer in beta.

---- Initial feature release ----

6) Fix discovered bugs.

7) Implement worthwhile user feedback if I feel like it and have time.