beta: Finish implementing makeWidget. (Excepting '--option')
All that remains is implementing the '--option' widget-making code,
possibly dealing with a complication involving arguments in the
'station' command having the same name but different type as in other
commands, and implementing the whole running of the commands.

Implementing the running of commands is next on the list, the other two
can be worked on post-release of the beta.