beta: Implement runTD
'--options' is currently a text entry, will be changed in future to be
procedural generated like the TD arguments currently are.

'--credits' is also currently a text entry. TD allows a number followed
by K,M,G. Numeric entries don't allow any letters, text entries allow
all, so for now it's text.

The output still goes straight to the terminal, that's next on the list.