fix: Implement plugin options subwindow
Now you don't need to type in every option. Push the button, and a
window will pop up allowing you to simply select which options you'd
like.

Plugin options which require a value, such as the filename to test json
importing with in the case of the edapi plugin's 'test' option, will
show as a text field for typing that parameter in.