feat: Allow for capacities above 1500
See #89

There is no longer an upper limit, capacities >1500 will trigger a
warning, not an error.

This also forces jumps per hop to 1 and adds '--supply' and '--demand',
if not already provided, to ignore stations that do not have
supply/demand of at least 10*capacity. This has the effect of filtering
out many stations, with the effect of making it more likely it will find
a route that not only will fill the hold and also more likely the route
can be repeated multiple times.

1 hop with the above settings is about long enough to shop for, prepare,
and eat dinner.

2 hops is multiple days without '--loop'. With '--loop', it took ~150%
of the time as 1 hop to finish, reporting failure to find a route.

It is extremely unlikely that any station not only has a large quantity
of something another station needs a large quantity of, but also needs
needs a large quantity of something that same other station has a large
quantity of, so not finding a 2-hop loop route isn't surprising.

It might be possible to find a loop route of 3+ hops, this has not been
tested and is not recommended. A 3-hop loop route might be possible to
find reliably, user discretion is advised.

Doing an unsupported >1500 capacity trade with >3 hops will trigger a
special warning with instructions on how to SIGTERM using a keyboard.