fix: maxage errors if not set
`float(null)` doesn't work, obviously, make sure to only cast to float
is maxage has been set. Also make sure maxage is set before doing the
math to see if a station should be skipped.

fix: Make sure the cache is updated if a system, station, or commodity
was added to the DB.

fix: Use fdev_id as item_id when adding new commodity to the DB.

fix: Update the ui_order for commodities when a new one is added to the
DB.

refactor: Use a better method for removing spaces from a prices file.