chore: tox test environments needs to require the packages that tradedangerous is going to be depending on (#122)
The tox testenv's need to import the same set of packages as dependencies that trade is going to want. using "-r requirements/dev.txt" tells it to do this.

I also added "flake8" as an env so that you can use tox to get flake results with tox -e flake8; the value there is that you get apples-to-apples. I hate running flake in my ide / vim and then tox has a slightly different config :)