feat: introduce theming in tradeenv
two fold reasoning:
1- so people can opt out of any kind of decoration if it becomes inconvenient/problematic,
2- localization and style

In particular: the colors people want for light-vs-dark background terminals might vary, and this would make it easy to switch; second, this makes it possible to configure various strings to be localizable through the theme.

The demo-case included here is that when you use --color the word "NOTE" will be replaced with the "information-source" emoji and the word "WARNING" will be replaced with a warning-symbol emoji, the '#' for bugs will be replaced with a bug emoji.