perf: (eddblink) get the timestamp for all station commodities at once
Do one SQL query before processing the listings file to get every
timestamp, rather than doing a SQL query for the timestamp on each
individual listing as we come to it.

This cuts the number of SQL statements being made during processing --
and thus the time taken to process -- roughly in half.