fix: only try to publish from eyeonus repos
fixes #134

This makes the publish job dependent on being the eyeonus repository. It also moves the requirements files
into the top-level directory and organizes them into requirements, requirements-dev, and requirements-publish
which incrementally reference the previous requirements file to simplify them.

It's unusual for requirements*.txt to be in a subdirectory these days, and various tools are lazy about it,
which can make these chained-references problematic as some tools will intepret them relatively and some
won't; ergo having them in the tld together is less likely to suddenly not work one day.