Fixes incorrect locale-based parsing in eddblink
A recent change was made to support non-English locale users with the
eddblink date parsing functions, which relied upon locale-specific
parsing. Unfortunately the submitted solution for that issue made
further locale-specific assumptions that break in different ways.

This change changes the system to instead parse the dates in
a locale-agnostic manner.

Tested by running the eddblink plugin on my system and seeing it was no
longer broken.