feat: enable repeat http requests over a single session
If we need to make multiple requests to a single http server, we have to repeat the overhead of connection-tear up which
can be significant for a remote https connection. 'requests' solves for this by letting you create a 'Session' object
that uses http keep-alive to send followup requests over an existing connection.

Commiting this separately from changes that make use of it.