fix: fix slow station loading
array lookups are not cached, so for every station it was having to do two dictionary table lookups to find the list of possible type ids. every time I hit ctrl-c (out of 16) it as on isFleet = 'Y'...

It would actually be best to eliminate the named values entirely in favor of type_id in (24, 0) and type_id == 25 to avoid doing an additional value lookup - global comparison is faster in this scenario, but not of big enough value to lose the named constants now.