feature: Persist TradeDB data between instances
My original intention for TradeDB() was that it be used many times,
to pay the tear-up cost, but we are now paying that cost in the
actual database.

This change provides a temporary bridge that reduces startup time
by using python's built-in functionality for shelving objects to
storage and back ('pickle') to save the post-load state of the
System, Station, Category, and Item tables between loads. On an
older PC this reduces startup time by 8-15 seconds.

The persistence file is data/TradeDB.pj (for "pickle jar")

During import or if the database file is modified, the previous pickle is discarded.