feat: 'units' option for eddblink import (#244)
Somehow eddb data has lots of bogus listing entries with a supply price but a units value of 0, such
as 'thardgoid probe' appearing at stations that clearly do not sell it.

This change makes it so that running eddblink import with -O units will treat those entries
as having a supply-price of zero or demand-price of zero depending on which field it is.

This has the knock-on effect of hiding lines that just arent' present at a station, but this
may have the consequence of making it impossible to look up good buyers for rare items like
Guardian Orb etc, which will only appear in the list if you have them in inventory